.CrossPlatformInput ERRORS FIXED
This isn't part of the course. But this file is one I had to search for because it was near the top of the list in the asset pack. So when I followed the instructions in the video this file didn't get included and lots of errors appeared.
Solution reimport this file from the standard assets.